Set scrollview on the power menu
So why? Because fuck you that's why...

No, you need this for if and when we decide to add more items to the power menu and the
density is too high. Previously if you had more than 5 items, it would cut you off. So
you either had to decide which 5 items you wanted or deal with the jank. That's no longer
the case.

- Added a landscape view so we can set a horizontal scrollview

- Made the power menu dialog all one color. Josh and I talked about this and I previously
made the case to keep it the same but after thinking it over, it looks better all one color.

Change-Id: I8ec4b1a85994251126433cea0640e000af78c65d
Signed-off-by: VenkatVishalV <venkatvishal124@gmail.com>

base: add missing noDivider boolean

Signed-off-by: spezi77 <spezi7713@gmx.net>