Return null if a provider died before acquisition
In case of a race between appDiedLocked and getContentProvider, it was
possible for a holder with a null provider to be returned.  The caller
of getContentProvider would then typically try to instantiate the
provider locally, and fail with a security exception.

Instead of returning an invalid holder with a null provider, return a
null holder.  getContentProvider already returns null in several cases,
so callers should be prepared for it.

Bug: 110800239
Test: manual
Change-Id: I70920e666fb46bdedec3592ea487018b9570a31f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>