Screenrecord service improvements and QSTile [1/2]
single tap to toggle quality:
- low quality resolution (480p) and 1.5Mbps
- medium resolution (720p) and 4Mbps (default)
- high quality resolution (1080p) and 8Mbps
long press to start recording

Show also a blinking red button at bottom right of the screen
that can be tapped to stop the recording instead of using
the notification.

Also improve the screenrecord service and be sure
to stop it when we are done.

part 2/2 is in framework/av tree, it just sets max recording
duration to 30 minutes instead of 3

Signed-off-by: spezi77 <spezi7713@gmx.net>